feat(settings): signed session cookies, but allow Lax
With 'Strict', the user loses their Django session between leaving the app
to sign in with OAuth, and coming back into the app from the OAuth redirect.

This is because Strict disallows our cookie being sent from an external
domain, and so the session cookie is lost on the return trip.