refactor: remove intermediary statuses
the thought at one point was that they would help drive the routing
between steps, but this turned out to be the responsibility of the views.

removing them means that once the status has been set to "submitted" it won't be moved back, which is necessary for the "already submitted" logic.