fix(ci): use the ubuntu-22.04 runner
ubuntu-latest went from 22.04 to 24.04 in the last few weeks
and we've seen about 30x slowdown in build times since then

see this run (on 22.04) taking ~5 mins:
https://github.com/Office-of-Digital-Services/cdt-ods-disaster-recovery/actions/runs/16950235801/job/48041152955

vs. this run (on 24.04) taking ~30 mins:
https://github.com/Office-of-Digital-Services/cdt-ods-disaster-recovery/actions/runs/17076099728/job/48417876227

this is attributable to building in emulation for the arm64 platform,
where the 22.04 toolchain seemed to handle this fine, and something in
24.04 causes a massive drop in performance of the emulated build steps

we'll pin to 22.04 for a while and keep an eye on announcements from
GitHub etc. as to when we might want to bump back up to 24.04 (or latest)