fix: override TypeForm.save method so it clears out fields appropriately
if the user has gone through the flow for a certain record type and then goes
back and changes the type, it is not desired for the previous values
to show up in the form fields.

for example, say a user started a request for a death record and filled out
"Parent's name" and "Spouse's name". If they go back and change their request
to be for a marriage record, we wouldn't want those values to show up as
"First person" and "Second person" on the marriage name form.