chore: move form_helpers to core app
since the custom template tag set could be used on different forms.
Even though form_helpers can be used by any app since Django will
find the form_helpers.py file automatically as long as the app is
registered, it's better to have it in the core app for clarity.